Add Sass stylesheets
As the stylesheet is growing in size, it is becoming harder to navigate
it. One solution to  would be to divide style.css into several smaller
files, but this would come with the cost of decreased performance due to
more HTTP requests.

The introduction of Sass would solve this as all of the stylesheets are
compiled to a single CSS file (and it would also come with other perks!)

Thus, Sass code was added. In the process, the styling was also
optimized.